backend/requirements.txt
chardet>=5.0,<6.0
fastapi==0.110.0
uvicorn[standard]==0.27.1
httpx>=0.24,<0.28
python-dotenv==1.0.1
pydantic>=2.6,<2.8
pydantic-settings>=2.2,<2.5
supabase==2.7.4
openai==1.14.0
python-pptx==0.6.23
reportlab==4.1.0
matplotlib==3.8.3
cryptography==42.0.5
python-multipart==0.0.9
pillow==10.2.0
facebook-business==19.0.0
google-analytics-data==0.18.5
google-ads==24.1.0
google-auth==2.28.1
google-auth-httplib2==0.2.0
resend==0.8.0
stripe==8.4.0
razorpay==1.4.2